ch2-10. private folder (_폴더)
(logout)/@modal/(.)i/flow/login/page.tsx 파일과
(logout)/i/flow/login/page.tsx 파일은 동일하기 때문에 컴포넌트로 만들어서 정리하면 좋은데
이를 위해 폴더명 맨 앞에 언더바를 사용한 프라이빗폴더를 활용하면 좋다.

이러한 프라이빗폴더 또한 브라우저 주소창에 뜨지 않는데, 주소창에 뜨지 않는 폴더 3개를 정리하자면
1. 폴더명을 소괄호로 감싼 그룹폴더는 레이아웃을 구분하기 위해 사용하고
2. 폴더명 맨앞에 @를 붙인 패러렐라우트 폴더는 한 화면에 두 개의 페이지를 동시에 보여줄 때 사용되고
3. 폴더명 맨앞에 언더바를 사용한 프라이빗폴더는 컴포넌트를 정리할 때처럼 단순히 파일을 정리할 때 사용한다.

그래서 공통적으로 사용하는 파일은 공통적인 레이아웃 그룹폴더 내부에 _component 이름의 프라이빗폴더를 만들고,
기존의 i/flow/login/page.tsx 파일에선 해당 공통컴포넌트를 단순히 가져와서 렌더링하도록 수정하면 된다.

이렇게 page.tsx 파일을 수정하면서 기존에 최상단에 use client 선언문을 통해 클라이언트 컴포넌트로 바뀌었지만, use client 선언문을 사용하지 않게 되면서 다시 서버 컴포넌트가 되었고,

이처럼 서버컴포넌트는 클라이언트 컴포넌트를 가져올 수 있지만, 클라이언트 컴포넌트는 서버 컴포넌트를 가져오게 되면 가져오는 서버 컴포넌트가 클라이언트 컴포넌트로 바뀌는 문제가 발생하기 때문에, 클라이언트 컴포넌트는 서버 컴포넌트를 가져오면 안된다.
